Rename tifscrollformview to tifscrollview to encompass scroll context and keyboard avoiding view behaviors Extract the scroll context to the tifscrollview Refactored parallax factor to a const instead of a prop Renamed mapsnippet to mappreview Updated map preview in storybook
