+++ .storybook/components/MapSnippet/MapSnippet.stories.tsx
@@ -1,12 +1,12 @@
-import { ExpandableMapSnippetView } from "@components/MapSnippetView"
-import { Headline } from "@components/Text"
-import { TiFFormScrollView } from "@components/form-components/ScrollView"
+import { MapPreview } from "@components/MapPreview"
+import { Headline, Title } from "@components/Text"
+import { TiFScrollView } from "@components/common/ScrollView"
 import { XEROX_ALTO_DEFAULT_REGION } from "@explore-events-boundary"
+import { faker } from "@faker-js/faker"
+import { PortalProvider } from "@gorhom/portal"
 import React, { useState } from "react"
-import { View, Text } from "react-native"
 import { SafeAreaProvider, SafeAreaView } from "react-native-safe-area-context"
 import { StoryMeta } from "storybook/HelperTypes"
-import { PortalProvider } from "@gorhom/portal"
 
 export const MapSnippetMeta: StoryMeta = {
   title: "MapSnippet"
@@ -14,22 +14,25 @@ export const MapSnippetMeta: StoryMeta = {
 
 export default MapSnippetMeta
 
+const firstContent = faker.lorem.paragraph(8)
+const secondContent = faker.lorem.paragraph(15)
+
 export const Basic = () => {
   const [isExpanded, setIsExpanded] = useState(false)
   return (
     <SafeAreaProvider>
       <PortalProvider>
         <SafeAreaView>
-          <TiFFormScrollView>
-            <View style={{ height: 64 }} />
-            <Headline>Hello World</Headline>
-            <ExpandableMapSnippetView
+          <TiFScrollView>
+            <Title>Map Snippet Demo</Title>
+            <Headline>{firstContent}</Headline>
+            <MapPreview
               isExpanded={isExpanded}
               onExpansionChanged={setIsExpanded}
               region={XEROX_ALTO_DEFAULT_REGION}
             />
-            <Headline>Hello World</Headline>
-          </TiFFormScrollView>
+            <Headline>{secondContent}</Headline>
+          </TiFScrollView>
         </SafeAreaView>
       </PortalProvider>
     </SafeAreaProvider>
